requirements.txt
# Flask==3.0.0
# gunicorn==20.1.0

# # YOLOv5 requirements
# matplotlib>=3.2.2
# numpy>=1.18.5
# opencv-python>=4.1.1
# Pillow>=7.1.2
# PyYAML>=5.3.1
# requests>=2.23.0
# scipy>=1.4.1
# torch>=1.7.0
# torchvision>=0.8.1
# tqdm>=4.64.0
# seaborn>=0.11.0
# pandas

# YOLOv5 optional requirements



Flask==2.2.2
gunicorn==19.9.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow ==9.1.0
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
tensorboard>=2.4.1
seaborn>=0.11.0
pandas>=0.19
thop == 0.0.31-2005241907
Werkzeug==2.3.7